+++ packages/api-client/src/api/addBundleProductsToCart/addBundleProductsToCart.ts
@@ -1,6 +1,4 @@
-import gql from "graphql-tag";
-
-export default gql`
+export default `
   mutation addBundleProductsToCart($input: AddBundleProductsToCartInput) {
     addBundleProductsToCart(input: $input) {
       cart {
@@ -11,26 +9,35 @@ export default gql`
           code
         }
         prices {
+          subtotal_with_discount_excluding_tax {
+            value
+            currency
+          }
           subtotal_excluding_tax {
             value
+            currency
           }
           subtotal_including_tax {
             value
+            currency
           }
           applied_taxes {
             amount {
               value
+              currency
             }
             label
           }
           discounts {
             amount {
               value
+              currency
             }
             label
           }
           grand_total {
             value
+            currency
           }
         }
         items {
@@ -80,6 +87,8 @@ export default gql`
               name
               url_suffix
               url_path
+              url_key
+              include_in_menu
               breadcrumbs {
                 category_name
                 category_url_path
@@ -99,12 +108,15 @@ export default gql`
           prices {
             row_total {
               value
+              currency
             }
             row_total_including_tax {
               value
+              currency
             }
             total_item_discount {
               value
+              currency
             }
           }
           quantity
@@ -117,6 +129,20 @@ export default gql`
             }
             configured_variant {
               sku
+              name
+              only_x_left_in_stock
+              price_range {
+                minimum_price {
+                  final_price {
+                    currency
+                    value
+                  }
+                  regular_price {
+                    currency
+                    value
+                  }
+                }
+              }
               thumbnail {
                 url
               }