+++ packages/api-client/src/api/addConfigurableProductsToCart/addConfigurableProductsToCart.ts
@@ -1,7 +1,5 @@
-import gql from "graphql-tag";
-
 /** GraphQL Mutation that adds configurable products to shopping cart */
-export default gql`
+export default `
   mutation addConfigurableProductsToCart($input: AddConfigurableProductsToCartInput) {
     addConfigurableProductsToCart(input: $input) {
       cart {
@@ -12,26 +10,35 @@ export default gql`
           code
         }
         prices {
+          subtotal_with_discount_excluding_tax {
+            value
+            currency
+          }
           subtotal_excluding_tax {
             value
+            currency
           }
           subtotal_including_tax {
             value
+            currency
           }
           applied_taxes {
             amount {
               value
+              currency
             }
             label
           }
           discounts {
             amount {
               value
+              currency
             }
             label
           }
           grand_total {
             value
+            currency
           }
         }
         items {
@@ -81,6 +88,8 @@ export default gql`
               name
               url_suffix
               url_path
+              url_key
+              include_in_menu
               breadcrumbs {
                 category_name
                 category_url_path
@@ -100,12 +109,15 @@ export default gql`
           prices {
             row_total {
               value
+              currency
             }
             row_total_including_tax {
               value
+              currency
             }
             total_item_discount {
               value
+              currency
             }
           }
           quantity
@@ -118,6 +130,20 @@ export default gql`
             }
             configured_variant {
               sku
+              name
+              only_x_left_in_stock
+              price_range {
+                minimum_price {
+                  final_price {
+                    currency
+                    value
+                  }
+                  regular_price {
+                    currency
+                    value
+                  }
+                }
+              }
               thumbnail {
                 url
               }