+++ packages/api-client/src/api/cart/cart.ts
@@ -9,26 +9,35 @@ export default `
         code
       }
       prices {
+        subtotal_with_discount_excluding_tax {
+          value
+          currency
+        }
         subtotal_excluding_tax {
           value
-        },
+          currency
+        }
         subtotal_including_tax {
           value
-        },
+          currency
+        }
         applied_taxes {
           amount {
             value
-          },
+            currency
+          }
           label
         }
         discounts {
           amount {
             value
-          },
+            currency
+          }
           label
         }
         grand_total {
           value
+          currency
         }
       }
       items {
@@ -78,8 +87,10 @@ export default `
             name
             url_suffix
             url_path
+            url_key
+            include_in_menu
             breadcrumbs {
-              category_name,
+              category_name
               category_url_path
             }
           }
@@ -97,12 +108,15 @@ export default `
         prices {
           row_total {
             value
+            currency
           }
           row_total_including_tax {
             value
+            currency
           }
           total_item_discount {
             value
+            currency
           }
         }
         quantity
@@ -115,6 +129,20 @@ export default `
           }
           configured_variant {
             sku
+            name
+            only_x_left_in_stock
+            price_range {
+              minimum_price {
+                final_price {
+                  currency
+                  value
+                }
+                regular_price {
+                  currency
+                  value
+                }
+              }
+            }
             thumbnail {
               url
             }