+++ packages/api-client/src/types/context/index.ts
@@ -2,4 +2,4 @@ import { type IntegrationContext } from "@vue-storefront/middleware";
 import { type ClientInstance, type Config } from "../setup";
 import { type Endpoints } from "../api";
 
-export type Context = IntegrationContext<ClientInstance, Config, Endpoints>;
+export interface Context extends IntegrationContext<ClientInstance, Config, Endpoints> {}