+++ packages/sdk/src/methods/index.ts
@@ -13,6 +13,8 @@ export * from './createEmptyCart';
 export * from './setShippingAddressesOnCart';
 export * from './removeCouponFromCart';
 export * from './addProductsToCart';
+export * from './addConfigurableProductsToCart';
+export * from './addBundleProductsToCart';
 export * from './setBillingAddressOnCart';
 export * from './setShippingMethodsOnCart';
 export * from './removeItemFromCart';